+++ src/taskpane/config.js
@@ -1,0 +1,5 @@
+const config = {
+    backendURL: "https://simplify-docs.appspot.com"
+}
+
+export default config
